FE: Move icons to public
