+++ frontend/app/(auth)/login/LoginForm.tsx
@@ -3,7 +3,7 @@ import { Button } from "@nextui-org/button";
 import { Input } from "@nextui-org/input"
 import  useAuth  from "../hooks/useAuth"
 import { Divider } from "@nextui-org/react";
-import { GithubIcon } from "../assets/GithubIcon";
+import { GithubIcon } from "../../../public/GithubIcon";
 
 export default function LoginForm() {
     const { register, errors, handleLogin, handleGithubLogin } = useAuth();
