+++ frontend/public/DeleteIcon.tsx
