+++ frontend/public/EyeIcon.tsx
